// src/app/src/app/enums/filters.enum.ts
export enum Filters {
  categories = 'categories',
  countries = 'countries',
  farmers = 'farmers',
  farms = 'farms',
  seals = 'seals',
  subcategories = 'subcategories'
}

export enum TrackingImpressionIds {
  FARMERS_MARKET_ADOPTIONS = 'Farmers Market Adoptions',
  FARMERS_MARKET_OH = 'Farmers Market OH',
  SEARCH_RESULTS_ADOPT = 'Search Results Adopt',
  SEARCH_RESULTS_OH = 'Search Results OH',
  SEARCH_RESULTS_ALL = 'Search Results All',
}

export enum TrackingGA4ImpressionIds {
  FARMERS_MARKET_ADOPTIONS = 'Farmers_Market_Adoptions',
  FARMERS_MARKET_OH = 'Farmers_Market_OH',
  SEARCH_RESULTS_ADOPT = 'Search_Results/Adoptions',
  SEARCH_RESULTS_OH = 'Search_Results/Boxes',
  SEARCH_RESULTS_ALL = 'Search_Results/All',
  SEARCH_NO_RESULTS = 'No_Results',
  SEARCH_RESULTS_ALL_ADOPTIONS_SUFFIX = '/Adoptions',
  SEARCH_RESULTS_ALL_OH_SUFFIX = '/Boxes',
  DIRECT = 'Direct',
}